+++ src/utility.h
@@ -3,14 +3,14 @@
 
 
 typedef struct vector2 {
-    int x;
-    int y;
+    float x;
+    float y;
 } vector2;
 
 typedef struct vector3 {
-    int x;
-    int y;
-    int z;
+    float x;
+    float y;
+    float z;
 } vector3;
 
 #define Vector2(x, y) (vector2){x, y}
